+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.test.ts
@@ -11,6 +11,7 @@ import { Field } from '@grafana/data';
 import {
   QuerySteps,
   QueryStepsResponse,
+  StepData,
   StepsProperties,
   StepsPropertiesOptions,
 } from 'datasources/results/types/QuerySteps.types';
@@ -23,7 +24,10 @@ import { ResultsDataSourceBase } from 'datasources/results/ResultsDataSourceBase
 import { Workspace } from 'core/types';
 import { DataSourceBase } from 'core/DataSourceBase';
 
-const mockQueryStepsResponse: QueryStepsResponse = {
+const mockSteps = Array(1000).fill({ stepId: '1', name: 'Step 1' });
+const mockPaths = Array(1000).fill({ path: 'path1' });
+
+const mockQueryStepsResponse = {
   steps: [
     {
       stepId: '1',
@@ -32,10 +36,10 @@ const mockQueryStepsResponse: QueryStepsResponse = {
         key1: 'value1',
         key2: 'value2',
       },
-      workspace: '1'
+      workspace: '1',
     },
   ],
-  continuationToken: undefined,
+  continuationToken: null,
   totalCount: 1,
 };
 
@@ -71,7 +75,6 @@ describe('QueryStepsDataSource', () => {
         createFetchResponse({
           paths: ['path1', 'path2'],
           continuationToken: null,
-          totalCount: 2,
         })
       );
   });
@@ -295,7 +298,6 @@ describe('QueryStepsDataSource', () => {
           createFetchResponse({
             steps: [],
             continuationToken: null,
-            totalCount: 0,
           } as unknown as QueryStepsResponse)
         );
 
@@ -346,13 +348,11 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch
         .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
         .mockReturnValue(createFetchResponse(mockQueryStepsResponse));
-        const query = buildQuery(
-          {
-            refId: 'A',
-            outputType: OutputType.Data,
-            properties: [StepsProperties.workspace]
-          },
-        );
+      const query = buildQuery({
+        refId: 'A',
+        outputType: OutputType.Data,
+        properties: [StepsProperties.workspace],
+      });
 
       const response = await datastore.query(query);
 
@@ -368,13 +368,11 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch
         .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
         .mockReturnValue(createFetchResponse(mockQueryStepsResponse));
-      const query = buildQuery(
-        {
-          refId: 'A',
-          outputType: OutputType.Data,
-          properties: [StepsProperties.workspace]
-        },
-      );
+      const query = buildQuery({
+        refId: 'A',
+        outputType: OutputType.Data,
+        properties: [StepsProperties.workspace],
+      });
 
       const response = await datastore.query(query);
 
@@ -382,11 +380,10 @@ describe('QueryStepsDataSource', () => {
       expect(fields).toMatchSnapshot();
     });
 
-
     describe('show measurements is enabled', () => {
       describe('duplicate measurement names', () => {
         beforeEach(() => {
-          const mockQueryStepsMeasurementResponse: QueryStepsResponse = {
+          const mockQueryStepsMeasurementResponse = {
             steps: [
               {
                 stepId: '1',
@@ -424,7 +421,7 @@ describe('QueryStepsDataSource', () => {
                 },
               },
             ],
-            totalCount: 1,
+            continuationToken: null,
           };
 
           backendServer.fetch
@@ -462,7 +459,7 @@ describe('QueryStepsDataSource', () => {
       });
 
       test('should create empty cells when measurements are not available', async () => {
-        const mockQueryStepsMeasurementResponse: QueryStepsResponse = {
+        const mockQueryStepsMeasurementResponse = {
           steps: [
             {
               stepId: '1',
@@ -479,7 +476,7 @@ describe('QueryStepsDataSource', () => {
               },
             },
           ],
-          totalCount: 1,
+          continuationToken: null,
         };
 
         backendServer.fetch
@@ -500,7 +497,7 @@ describe('QueryStepsDataSource', () => {
       });
 
       test('should create new columns when units are different in the same measurement', async () => {
-        const mockQueryStepsMeasurementResponse: QueryStepsResponse = {
+        const mockQueryStepsMeasurementResponse = {
           steps: [
             {
               stepId: '1',
@@ -517,7 +514,7 @@ describe('QueryStepsDataSource', () => {
               },
             },
           ],
-          totalCount: 1,
+          continuationToken: null,
         };
 
         backendServer.fetch
@@ -550,7 +547,7 @@ describe('QueryStepsDataSource', () => {
           parameters: [{ name: 'Voltage', measurement: '3.7', unit: 'V' }],
         };
 
-        const mockQueryStepsConditionsResponse: QueryStepsResponse = {
+        const mockQueryStepsConditionsResponse = {
           steps: [
             {
               ...mockQueryStepsResponse.steps[0],
@@ -567,7 +564,7 @@ describe('QueryStepsDataSource', () => {
               data: mockMeasurementData,
             },
           ],
-          totalCount: 2,
+          continuationToken: null,
         };
 
         backendServer.fetch
@@ -649,7 +646,7 @@ describe('QueryStepsDataSource', () => {
           ],
         };
 
-        const mockQueryStepsConditionsResponse: QueryStepsResponse = {
+        const mockQueryStepsConditionsResponse = {
           steps: [
             {
               ...mockQueryStepsResponse.steps[0],
@@ -666,7 +663,7 @@ describe('QueryStepsDataSource', () => {
               data: mockMeasurementData,
             },
           ],
-          totalCount: 2,
+          continuationToken: null,
         };
 
         backendServer.fetch
@@ -734,7 +731,7 @@ describe('QueryStepsDataSource', () => {
         ],
       };
 
-      const mockQueryStepsConditionsResponse: QueryStepsResponse = {
+      const mockQueryStepsConditionsResponse = {
         steps: [
           {
             ...mockQueryStepsResponse.steps[0],
@@ -751,7 +748,7 @@ describe('QueryStepsDataSource', () => {
             data: mockMeasurementData,
           },
         ],
-        totalCount: 2,
+        continuationToken: null,
       };
 
       backendServer.fetch
@@ -806,7 +803,6 @@ describe('QueryStepsDataSource', () => {
             },
           ],
           continuationToken: null,
-          totalCount: 1,
         } as unknown as QueryStepsResponse)
       );
 
@@ -824,12 +820,12 @@ describe('QueryStepsDataSource', () => {
     test('should call query steps API once when output type is total count', async () => {
       const mockResponses = [
         createFetchResponse({
-          steps: Array(100).fill({ stepId: '1', name: 'Step 1' }),
+          steps: mockSteps.slice(0, 100),
           continuationToken: null,
           totalCount: 5000,
         }),
         createFetchResponse({
-          steps: Array(100).fill({ stepId: '2', name: 'Step 2' }),
+          steps: mockSteps.slice(0, 100),
           continuationToken: null,
           totalCount: 5000,
         }),
@@ -935,21 +931,12 @@ describe('QueryStepsDataSource', () => {
     it('should make a single request when take is less than MAX_TAKE_PER_REQUEST', async () => {
       const mockResponses = [
         createFetchResponse({
-          steps: Array(100).fill({ stepId: '1', name: 'Step 1' }),
+          steps: mockSteps.slice(0, 100),
           continuationToken: null,
-          totalCount: 100,
         }),
       ];
       backendServer.fetch.mockImplementationOnce(() => mockResponses[0]);
-      const responsePromise = datastore.queryStepsInBatches(
-        undefined,
-        undefined,
-        undefined,
-        100,
-        undefined,
-        undefined,
-        true
-      );
+      const responsePromise = datastore.queryStepsInBatches(undefined, undefined, undefined, 100, undefined, undefined);
       const response = await responsePromise;
 
       expect(response.steps).toHaveLength(100);
@@ -962,17 +949,15 @@ describe('QueryStepsDataSource', () => {
       );
     });
 
-    test('should batch requests when total number od steps matching the filter is less than requested take', async () => {
+    test('should batch requests when total number of steps matching the filter is less than requested take', async () => {
       const mockResponses = [
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '1', name: 'Step 1' }),
+          steps: mockSteps.slice(0, 500),
           continuationToken: 'token1',
-          totalCount: 900,
         }),
         createFetchResponse({
-          steps: Array(400).fill({ stepId: '1', name: 'Step 1' }),
-          continuationToken: 'token2',
-          totalCount: 900,
+          steps: mockSteps.slice(0, 400),
+          continuationToken: null,
         }),
       ];
       backendServer.fetch.mockImplementationOnce(() => mockResponses[0]).mockImplementationOnce(() => mockResponses[1]);
@@ -982,8 +967,7 @@ describe('QueryStepsDataSource', () => {
         undefined,
         10000,
         undefined,
-        undefined,
-        true
+        undefined
       );
       const response = await responsePromise;
 
@@ -998,43 +982,34 @@ describe('QueryStepsDataSource', () => {
       expect(backendServer.fetch).toHaveBeenNthCalledWith(
         2,
         expect.objectContaining({
-          data: expect.objectContaining({ take: 400, continuationToken: 'token1' }),
+          data: expect.objectContaining({ take: 1000, continuationToken: 'token1' }),
         })
       );
     });
 
-    test('should batch requests with RequestPerSecond', async () => {
+    test('should batch requests and expand max take from 500 to 1000', async () => {
       jest.useFakeTimers();
       const fetchSpy = jest.spyOn(backendServer, 'fetch');
 
       const mockResponses = [
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '1', name: 'Step 1' }),
+          steps: mockSteps.slice(0, 500),
           continuationToken: 'token1',
-          totalCount: 2000,
         }),
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '2', name: 'Step 2' }),
+          steps: mockSteps,
           continuationToken: 'token2',
-          totalCount: 2000,
         }),
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '3', name: 'Step 3' }),
-          continuationToken: 'token3',
-          totalCount: 2000,
-        }),
-        createFetchResponse({
-          steps: Array(500).fill({ stepId: '4', name: 'Step 4' }),
+          steps: mockSteps.slice(0, 500),
           continuationToken: null,
-          totalCount: 2000,
         }),
       ];
 
       backendServer.fetch
         .mockImplementationOnce(() => mockResponses[0])
         .mockImplementationOnce(() => mockResponses[1])
-        .mockImplementationOnce(() => mockResponses[2])
-        .mockImplementationOnce(() => mockResponses[3]);
+        .mockImplementationOnce(() => mockResponses[2]);
 
       const responsePromise = datastore.queryStepsInBatches(
         undefined,
@@ -1042,14 +1017,82 @@ describe('QueryStepsDataSource', () => {
         undefined,
         2000,
         undefined,
-        undefined,
-        true
+        undefined
       );
 
       await jest.advanceTimersByTimeAsync(0);
       expect(fetchSpy).toHaveBeenCalledTimes(2);
       await jest.advanceTimersByTimeAsync(1000);
-      expect(fetchSpy).toHaveBeenCalledTimes(4);
+      expect(fetchSpy).toHaveBeenCalledTimes(3);
+      await responsePromise;
+
+      expect(fetchSpy).toHaveBeenNthCalledWith(
+        1,
+        expect.objectContaining({
+          data: expect.objectContaining({ take: 500, continuationToken: undefined }),
+        })
+      );
+      expect(fetchSpy).toHaveBeenNthCalledWith(
+        2,
+        expect.objectContaining({
+          data: expect.objectContaining({ take: 1000, continuationToken: 'token1' }),
+        })
+      );
+      expect(fetchSpy).toHaveBeenNthCalledWith(
+        3,
+        expect.objectContaining({
+          data: expect.objectContaining({ take: 500, continuationToken: 'token2' }),
+        })
+      );
+    });
+
+    test('should batch requests and continue with take of 500 when number of measurements is greater than 25', async () => {
+      jest.useFakeTimers();
+      const fetchSpy = jest.spyOn(backendServer, 'fetch');
+
+      const measurements: StepData = {
+        text: 'Step with many measurements',
+        parameters: Array(30).fill({ name: `Measurement`, measurement: '1.0' }),
+      };
+
+      const stepsWithManyMeasurements = mockSteps.map(step => ({
+        ...step,
+        data: measurements,
+      }));
+
+      const mockResponses = [
+        createFetchResponse({
+          steps: stepsWithManyMeasurements.slice(0, 500),
+          continuationToken: 'token1',
+        }),
+        createFetchResponse({
+          steps: stepsWithManyMeasurements.slice(0, 500),
+          continuationToken: 'token2',
+        }),
+        createFetchResponse({
+          steps: stepsWithManyMeasurements.slice(0, 500),
+          continuationToken: null,
+        }),
+      ];
+
+      backendServer.fetch
+        .mockImplementationOnce(() => mockResponses[0])
+        .mockImplementationOnce(() => mockResponses[1])
+        .mockImplementationOnce(() => mockResponses[2]);
+
+      const responsePromise = datastore.queryStepsInBatches(
+        undefined,
+        undefined,
+        undefined,
+        1500,
+        undefined,
+        undefined
+      );
+
+      await jest.advanceTimersByTimeAsync(0);
+      expect(fetchSpy).toHaveBeenCalledTimes(2);
+      await jest.advanceTimersByTimeAsync(1000);
+      expect(fetchSpy).toHaveBeenCalledTimes(3);
       await responsePromise;
 
       expect(fetchSpy).toHaveBeenNthCalledWith(
@@ -1070,33 +1113,19 @@ describe('QueryStepsDataSource', () => {
           data: expect.objectContaining({ take: 500, continuationToken: 'token2' }),
         })
       );
-      expect(fetchSpy).toHaveBeenNthCalledWith(
-        4,
-        expect.objectContaining({
-          data: expect.objectContaining({ take: 500, continuationToken: 'token3' }),
-        })
-      );
     });
 
     test('should stop fetching when continuationToken is null', async () => {
       const mockResponses = [
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '1', name: 'Step 1' }),
+          steps: mockSteps.slice(0, 500),
           continuationToken: null,
           totalCount: 500,
         }),
       ];
 
       backendServer.fetch.mockImplementationOnce(() => mockResponses[0]);
-      const response = await datastore.queryStepsInBatches(
-        undefined,
-        undefined,
-        undefined,
-        3000,
-        undefined,
-        undefined,
-        true
-      );
+      const response = await datastore.queryStepsInBatches(undefined, undefined, undefined, 3000, undefined, undefined);
 
       expect(response.steps).toHaveLength(500);
       expect(backendServer.fetch).toHaveBeenCalledTimes(1);
@@ -1111,21 +1140,19 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch
         .mockImplementationOnce(() =>
           createFetchResponse({
-            steps: Array(500).fill({ stepId: '1', name: 'Step 1' }),
+            steps: mockSteps.slice(0, 500),
             continuationToken: 'token1',
-            totalCount: 1500,
           })
         )
         .mockImplementationOnce(() => createFetchError(400)) //Error
         .mockImplementationOnce(() =>
           createFetchResponse({
-            steps: Array(500).fill({ stepId: '2', name: 'Step 2' }),
+            steps: mockSteps.slice(0, 500),
             continuationToken: 'token2',
-            totalCount: 1500,
           })
         );
 
-      const batchPromise = datastore.queryStepsInBatches('filter', 'orderBy', undefined, 1500, false, undefined, true);
+      const batchPromise = datastore.queryStepsInBatches('filter', 'orderBy', undefined, 1500, false, undefined);
 
       await expect(batchPromise).rejects.toThrow('The query failed due to the following error: (status 400) "Error".');
       expect(backendServer.fetch).toHaveBeenCalledTimes(2);
@@ -1143,46 +1170,36 @@ describe('QueryStepsDataSource', () => {
 
       const mockResponses = [
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '1', name: 'Step 1' }),
+          steps: mockSteps.slice(0, 500),
           continuationToken: 'token1',
-          totalCount: 2000,
         }),
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '2', name: 'Step 2' }),
+          steps: mockSteps,
           continuationToken: 'token2',
-          totalCount: 2000,
         }),
         createFetchResponse({
-          steps: Array(500).fill({ stepId: '3', name: 'Step 3' }),
+          steps: mockSteps.slice(0, 500),
           continuationToken: 'token3',
-          totalCount: 2000,
-        }),
-        createFetchResponse({
-          steps: Array(500).fill({ stepId: '4', name: 'Step 3' }),
-          continuationToken: null,
-          totalCount: 2000,
         }),
       ];
       backendServer.fetch
         .mockImplementationOnce(() => mockResponses[0])
         .mockImplementationOnce(() => mockResponses[1])
-        .mockImplementationOnce(() => mockResponses[2])
-        .mockImplementationOnce(() => mockResponses[3]);
+        .mockImplementationOnce(() => mockResponses[2]);
       const responsePromise = datastore.queryStepsInBatches(
         undefined,
         undefined,
         undefined,
         2000,
         undefined,
-        undefined,
-        true
+        undefined
       );
 
       const response = await responsePromise;
 
       expect(response.steps).toHaveLength(2000);
-      expect(backendServer.fetch).toHaveBeenCalledTimes(4);
-      expect(spyDelay).toHaveBeenCalledTimes(1);
+      expect(backendServer.fetch).toHaveBeenCalledTimes(3);
+      expect(spyDelay).toHaveBeenCalledTimes(2);
       expect(spyDelay).toHaveBeenCalledWith(expect.any(Function), 800); // delay for 1000 - 200 = 800ms
     });
   });
@@ -1191,9 +1208,8 @@ describe('QueryStepsDataSource', () => {
     it('should make a single request when take is less than MAX_PATH_TAKE_PER_REQUEST', async () => {
       const mockResponses = [
         createFetchResponse({
-          paths: Array(100).fill({ path: 'path1' }),
+          paths: mockPaths.slice(0, 100),
           continuationToken: null,
-          totalCount: 100,
         }),
       ];
       backendServer.fetch.mockImplementationOnce(() => mockResponses[0]);
@@ -1214,14 +1230,12 @@ describe('QueryStepsDataSource', () => {
     test('should batch requests when total number of paths matching the filter is less than requested take', async () => {
       const mockResponses = [
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path1' }),
+          paths: mockPaths,
           continuationToken: 'token1',
-          totalCount: 1500,
         }),
         createFetchResponse({
-          paths: Array(500).fill({ path: 'path2' }),
-          continuationToken: 'token2',
-          totalCount: 1500,
+          paths: mockPaths.slice(0, 500),
+          continuationToken: null,
         }),
       ];
       backendServer.fetch.mockImplementationOnce(() => mockResponses[0]).mockImplementationOnce(() => mockResponses[1]);
@@ -1241,7 +1255,7 @@ describe('QueryStepsDataSource', () => {
         2,
         expect.objectContaining({
           url: '/nitestmonitor/v2/query-paths',
-          data: expect.objectContaining({ take: 500, continuationToken: 'token1' }),
+          data: expect.objectContaining({ take: 1000, continuationToken: 'token1' }),
         })
       );
     });
@@ -1252,24 +1266,20 @@ describe('QueryStepsDataSource', () => {
 
       const mockResponses = [
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path1' }),
+          paths: mockPaths,
           continuationToken: 'token1',
-          totalCount: 4000,
         }),
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path2' }),
+          paths: mockPaths,
           continuationToken: 'token2',
-          totalCount: 4000,
         }),
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path3' }),
+          paths: mockPaths,
           continuationToken: 'token3',
-          totalCount: 4000,
         }),
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path4' }),
+          paths: mockPaths,
           continuationToken: null,
-          totalCount: 4000,
         }),
       ];
 
@@ -1320,9 +1330,8 @@ describe('QueryStepsDataSource', () => {
     test('should stop fetching when continuationToken is null', async () => {
       const mockResponses = [
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path1' }),
+          paths: mockPaths,
           continuationToken: null,
-          totalCount: 1000,
         }),
       ];
 
@@ -1343,17 +1352,15 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch
         .mockImplementationOnce(() =>
           createFetchResponse({
-            paths: Array(1000).fill({ path: 'path1' }),
+            paths: mockPaths,
             continuationToken: 'token1',
-            totalCount: 3000,
           })
         )
         .mockImplementationOnce(() => createFetchError(400)) //Error
         .mockImplementationOnce(() =>
           createFetchResponse({
-            paths: Array(1000).fill({ path: 'path2' }),
+            paths: mockPaths,
             continuationToken: 'token2',
-            totalCount: 3000,
           })
         );
 
@@ -1377,24 +1384,20 @@ describe('QueryStepsDataSource', () => {
 
       const mockResponses = [
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path1' }),
+          paths: mockPaths,
           continuationToken: 'token1',
-          totalCount: 4000,
         }),
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path2' }),
+          paths: mockPaths,
           continuationToken: 'token2',
-          totalCount: 4000,
         }),
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path3' }),
+          paths: mockPaths,
           continuationToken: 'token3',
-          totalCount: 4000,
         }),
         createFetchResponse({
-          paths: Array(1000).fill({ path: 'path4' }),
+          paths: mockPaths,
           continuationToken: null,
-          totalCount: 4000,
         }),
       ];
       backendServer.fetch
@@ -1437,7 +1440,6 @@ describe('QueryStepsDataSource', () => {
           createFetchResponse({
             paths: [{ path: 'path1' }, { path: 'path2' }, { path: 'path1' }, { path: 'path3' }, { path: 'path2' }],
             continuationToken: null,
-            totalCount: 5,
           })
         );
 
@@ -1652,8 +1654,8 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch.mockReturnValue(
         createFetchResponse({
           steps: [{ name: 'StepA' }, { name: 'StepB' }],
-          totalCount: 2,
-        } as QueryStepsResponse)
+          continuationToken: null,
+        } as unknown as QueryStepsResponse)
       );
 
       const query = { queryByResults: 'programName = "name"', stepsTake: 1000 } as StepsVariableQuery;
@@ -1669,8 +1671,8 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch.mockReturnValue(
         createFetchResponse({
           steps: [],
-          totalCount: 0,
-        } as QueryStepsResponse)
+          continuationToken: null,
+        } as unknown as QueryStepsResponse)
       );
 
       const query = { queryByResults: 'programName = "name"', stepsTake: 1000 } as StepsVariableQuery;
@@ -1709,8 +1711,8 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch.mockReturnValue(
         createFetchResponse({
           steps: [{ name: 'Step1' }],
-          totalCount: 1,
-        } as QueryStepsResponse)
+          continuationToken: null,
+        } as unknown as QueryStepsResponse)
       );
 
       const query = { queryByResults: resultsQuery, queryBySteps: stepsQuery, stepsTake: 1000 } as StepsVariableQuery;