+++ src/datasources/test-plans/TestPlansDataSource.test.ts
@@ -698,7 +698,6 @@ describe('runQuery', () => {
       ["ID"],
       1000,
       undefined,
-      true,
     );
   });
 
@@ -722,7 +721,6 @@ describe('runQuery', () => {
       ["ID"],
       1000,
       undefined,
-      true,
     );
   });
 
@@ -746,7 +744,6 @@ describe('runQuery', () => {
       ["ID"],
       1000,
       undefined,
-      true,
     );
 
     jest.useRealTimers();
@@ -822,7 +819,6 @@ describe('runQuery', () => {
       ["ID"],
       1000,
       undefined,
-      true
     );
   });
 
@@ -1166,7 +1162,6 @@ describe('queryTestPlansInBatches', () => {
         { id: '2', name: 'Test Plan 2' }
       ],
       continuationToken: undefined,
-      totalCount: 2,
     };
 
     jest.spyOn(datastore, 'queryTestPlans').mockResolvedValue(mockQueryResponse);
@@ -1174,7 +1169,6 @@ describe('queryTestPlansInBatches', () => {
     const result = await datastore.queryTestPlansInBatches('', OrderByOptions.UPDATED_AT, [Projections.NAME], 2, true);
 
     expect(result.testPlans).toEqual(mockQueryResponse.testPlans);
-    expect(result.totalCount).toEqual(2);
   });
 
   test('handles errors during batch querying', async () => {